% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{altitudes}
\alias{altitudes}
\title{Some craft and their altitudes}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{craft}{the name of the craft}
\item{altitude}{the craft's normal cruising altitude in metres}
}
}
\usage{
altitudes
}
\description{
Some craft and their altitudes
}
\keyword{datasets}
