% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals_rich}
\alias{animals_rich}
\title{Some animals and their attributes, in rich text}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{animal}{the name of the animal}, in rich text
\item{type}{the type of animal}
\item{flies}{whether the animal can fly}
\item{mass}{the average mass of the animal in grams}
}
}
\usage{
animals_rich
}
\description{
Some animals and their attributes, in rich text
}
\keyword{datasets}
