% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beverages}
\alias{beverages}
\title{The compositions of some espresso-based beverages}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{beverage}{the name of the beverage}
\item{ingredient}{the ingredient}
\item{proportion}{the proportion of a cup to fill with the ingredient}
}

A data frame with 6 rows and 3 variables:
\describe{
\item{beverage}{the name of the beverage}
\item{ingredient}{the ingredient, in rich text}
\item{proportion}{the proportion of a cup to fill with the ingredient}
}
}
\usage{
beverages

beverages
}
\description{
The compositions of some espresso-based beverages

The compositions of some espresso-based beverages, in rich text
}
\keyword{datasets}
