% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beverages_rich}
\alias{beverages_rich}
\title{The compositions of some espresso-based beverages, in rich text}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{beverage}{the name of the beverage}
\item{ingredient}{the ingredient, in rich text}
\item{proportion}{the proportion of a cup to fill with the ingredient}
}
}
\usage{
beverages_rich
}
\description{
The compositions of some espresso-based beverages, in rich text
}
\keyword{datasets}
