% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coffees}
\alias{coffees}
\title{The composition of two espresso-based beverages}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{coffee}{the name of the beverage}
\item{ingredient}{the ingredient}
\item{proportion}{the proportion of a cup to fill with the ingredient}
}
}
\usage{
coffees
}
\description{
The composition of two espresso-based beverages
}
\keyword{datasets}
