% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fill_text.R, R/geom_shrink_text.R
\name{geom_fill_text}
\alias{geom_fill_text}
\alias{geom_shrink_text}
\title{Fit text to a bounding box.}
\usage{
geom_fill_text(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, padding.x = unit(1, "mm"), padding.y = unit(0.1,
  "lines"), min.size = NULL, ...)

geom_shrink_text(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, padding.x = unit(1, "mm"), padding.y = unit(0.1,
  "lines"), place = "centre", min.size = NULL, ...)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{Standard
geom arguments as for ‘geom_text’. Note that x and y aesthetics will be
ignored; xmin, xmax, ymin and ymax aesthetics specifying the bounding box are
required.}

\item{padding.x}{Amount of padding around text horizontally, as a grid ‘unit’
object. Default is 1 mm.}

\item{padding.y}{Amount of padding around text vertically, as a grid ‘unit’
object. Default is 0.1 lines.}

\item{min.size}{Minimum font size. If specified, text that would need to be
shrunk below this size to fit the bounding box will not be drawn.}

\item{place}{Where to place the text within the bounding box. For
\code{geom_shrink_text}, default is ‘centre’, other options are ‘topleft’,
‘top‘, etc. For \code{geom_fill_text}, default is ‘middle’, other options are
‘top’ and ‘bottom’.}
}
\description{
\code{geom_shrink_text} will draw the label text normally, unless it is too
big for the bounding box, in which case it will shrink the text to fit the
box.

\code{geom_fill_text} will shrink or expand the label text to fill the box.

Except where noted, these geoms should behave like \code{geom_text}. In
addition to the normal \code{geom_text} aesthetics, \code{ggtextfit} geoms
use xmin, xmax, ymin and ymax to specify the bounding box for the label text
(x and y aesthetics, if passed, will be ignored).
}

