% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fit_text.R
\docType{data}
\name{GeomFitText}
\alias{GeomFitText}
\title{GeomFitText}
\format{
An object of class \code{GeomFitText} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomFitText
}
\description{
GeomFitText
}
\keyword{internal}
