% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfittext-package.R
\docType{package}
\name{ggfittext-package}
\alias{ggfittext}
\alias{ggfittext-package}
\title{ggfittext: Fit Text Inside a Box in 'ggplot2'}
\description{
A 'ggplot2' extension to fit text into a box by growing, shrinking or wrapping the text.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wilkox.org/ggfittext/}
  \item Report bugs at \url{https://github.com/wilkox/ggfittext/issues/}
}

}
\author{
\strong{Maintainer}: David Wilkins \email{david@wilkox.org}

Other contributors:
\itemize{
  \item Brady Johnston \email{brady.johnston@me.com} [contributor]
}

}
\keyword{internal}
