% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfittext.R
\docType{package}
\name{ggfittext}
\alias{ggfittext}
\alias{ggfittext-package}
\title{'ggfittext': provides a 'ggplot2' geom to fit text inside a box.}
\description{
This package provides only one function, \code{geom_fit_text}.
}
