% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gold}
\alias{gold}
\title{Robert Frost's poem \emph{Nothing Gold Can Stay}}
\format{
A data frame with 8 rows and 5 variables:
\describe{
\item{label}{a line from the poem}
\item{xmin}{the xmin coordinate with which to draw the word}
\item{xmax}{the xmax coordinate with which to draw the word}
\item{ymin}{the ymin coordinate with which to draw the word}
\item{ymax}{the ymax coordinate with which to draw the word}
\item{line}{the line number}
}
}
\usage{
gold
}
\description{
Robert Frost's poem \emph{Nothing Gold Can Stay}
}
\keyword{datasets}
