% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar_text.R, R/geom_fit_text.R
\docType{data}
\name{GeomBarText}
\alias{GeomBarText}
\alias{GeomFitText}
\title{GeomFitText}
\format{
An object of class \code{GeomBarText} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{GeomFitText} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomBarText

GeomFitText
}
\description{
GeomFitText
}
\keyword{internal}
